+++ src/cli/lint/internal.test.ts
@@ -1,0 +1,6 @@
+import { internalLint } from './internal';
+
+describe('internalLint', () => {
+  it('passes on skuba itself', () =>
+    expect(internalLint()).resolves.toBeUndefined());
+});
